feat: Add delete functionality for alerts
This commit adds the ability to delete alerts in the Home component. The `handleDelete` function is implemented to delete data from the "ignore-reminder" table in the Supabase database. After deleting the data, the `fetchData` function is called to update the UI. The delete functionality is triggered by clicking the "Delete" button in the alert section.

Refs: #123